+++ src/ovh-ui-angular/controllers/data/countries.json
