+++ src/ovh-ui-angular/controllers/select.controller.js
@@ -1,4 +1,4 @@
-import countries from "./countries.json";
+import countries from "./data/countries.json";
 
 export default class SelectController {
     constructor ($http, $timeout) {
